pyproject.toml
[project]
name = "queria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35",
    "duckdb>=1.4.0,<1.5.0",
    "jinja2>=3.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.15",
    "dbt-core>=1.11.2",
]

[project.optional-dependencies]
dlt = ["dlt[ducklake]>=1.6"]
metadata = ["dbt-core>=1.11.2"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/queria"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
docs = ["mkdocs-material>=9"]

[project.scripts]
queria = "queria.cli:main"
